improve(ollama): isolate lazy runtime modules
