+++ extensions/ollama/src/stream-registration.ts
@@ -1,7 +1,7 @@
 import type { StreamFn } from "openclaw/plugin-sdk/agent-core";
 import { createLazyRuntimeModule } from "openclaw/plugin-sdk/lazy-runtime";
 
-const loadOllamaStreamRuntime = createLazyRuntimeModule(() => import("./stream.js"));
+const loadOllamaStreamRuntime = createLazyRuntimeModule(() => import("./stream.runtime.js"));
 
 export function createLazyConfiguredOllamaStreamFn(params: {
   model: { baseUrl?: string; headers?: unknown };
